split txn tests
